fix scaling of I_PV* and I_AC
The modbus code was using the wrong scaling factor for input registers
I_PV1, I_PV2 and I_AC - should be DECI rather than CENTI, since the
raw register appears to be in units of 0.1A

givtcp was compensating by multiplying the I_PV values received by 10.
Better to fix it in the underlying library than in the client code.